Implemented a new method for retrieving the Subversion 1.7 revision number. This involves finding wc.db and using PDO's SQLite driver. This will not work if we can't find wc.db, and/or the PDO and/or pdo_sqlite extensions aren't loaded. In order to save time, the loop is skipped if the PDO extensions aren't loaded.
We can safely assume that if this is a Subversion 1.7 repository there will be a wc.db file somewhere in a previous directory. A $searchDepth of 3 will cover checkouts from /, /trunk, and /branches/branch. Thanks to some basename(realpath('..')) magic in the query SQL, it's able to automatically detect your branch name and get the revision number for that specific branch.

From my extremely limited testing the SQL query on line 466 isn't very fast. This is arguably more portable than the svnversion exec because it doesn't rely on external binaries. The primary barrier to entry will be people having (or not having) PDO and pdo_sqlite. In the future, as PDO use increases, this should become a non-issue.