Rewrote step 3 UI, removed old HTML
Step 3 is now home to "Basic Account Info." This includes common
information that is required for THT to process the account: username,
password, domain/subdomain, and email. Ajax calls have been implemented
on the client side to validate username and email fields. This also
needs to be done for the domain and, arugably, the password.

TODO: Optional captcha implementation. Use sidebar for account info
review