Package Manager rewrite, work in progress
This was long overdue and represents days of work. The old editor was
fragile and didn't even work correctly in some not-so-edge cases.
This is a single-page load-once AJAX solution. This is based off the
custom field code but heavily revised to be a tad more robust. There
are a number of issues with that old implementation that can be
fixed by porting this code back.

I actually rewrote the package manager a few years(?) back using the
jQuery UI but I looked through the old code and wasn't terribly
impressed. Ideally ajax.php should only be outputting JSON. It was
using ajax to actually output full blocks of HTML generated on the
server. The UI wouldn't scale as nicely either to accommodate extra
elements. The JS and HTML code is still available in the tpl
directory and corresponding AJAX code.

A new SQL field was added to the packages table to allow TLDs on a
per-package basis. Previously this was a global setting. The SQL for
the is_disabled field was already present. I must have been planning
on adding that functionality at some time before.

What works:
- The name, backend, description, type, server fields and all 5
checkboxes are wired up

- Editing a package using those fields

- Creating a new package using those fields

- Ordering packages

- Deleting packages

What doesn't:
- Editing type-specific info like with Post2Host

- Custom fields only display, they do not save or behave correctly.
  This includes data display. The fields are effectively static.

- Some weird formatting errors that may be the result of a mismatched
  HTML tag or CSS rule

- Cool CSS3 transition don't always work and the grayout of disabled
 packages only seems to work in WebKit and Gecko.

- The spinner only seems to display correctly in Chrome

- Internet Explorer has issues with the cursors

- No TinyMCE description editor yet

- qTip doesn't seem to work at all on IE11

- A metric shit-ton of hacky CSS styling (both inline and otherwise)
  That all needs to be organized and moved into a theme's style.css
  At present the default Reloaded2 theme is the only theme that this
  may actually work with.

- Info icons don't vertically align for the first five fields

- Probably other things.