Custom field support added to new package manager
Custom order fields have been reintroduced. Minimal testing but it
appears to work perfectly so far. The JS correctly handles new
packages with custom fields selected as well as existing packages.
Before this commit the custom field checkboxes would show but were
entirely non-functional. They are now fully functional.

Also changed the font to Lucida Grande / Sans.

A number of bugs still exist from the original package manager commit
however a large majority of them are visual in nature. Notably,
TinyMCE needs to be updated and implemented for package descriptions.

The moral of the story is the package manager works pretty damn well
in Chrome v36 dev for Mac. On Firefox? Eh. In Internet Explorer? Meh.
On Windows it's pretty much the same story. Except some rendering
issues cause things to not look quite as uniform.

CSS still needs to be organized and moved to style.css in addition
to the above remarks on compatibility.