Introduced grow/shrink ability to the pkg editor
This seriously helps low resolution screens handle the intense width
requirements of the package editor. It's also very useful on larger
screens to give the user a little more breathing room.

Functionally, the new package editor is nearing completion. The
majority of work now resides in making it styleable via user themes.
Ideally I'd like to make everything fit better when screen space is
limited, but that's not a blocking issues.

spin.js was also updated to v2.0.1